View
The view section is the section where I have an issue at the moment. The goal is for the user to input a number which is associated to a contact in the list. The program should then spit out the information associated with that specific number. Currently, the error is anything over 4 comes back with an error message.